requirements.txt
streamlit
google-generativeai
arxiv
pandas
python-dotenv
